bare rearrangement
